HisAppraiser: Fixed wrong checks in submitReport() function
This patch fixes two wrong checks. At first, the check on
'latestPolledRequest' variable is replaced with
'latestPolledRequest.getId()'. The reason why the original check
is wrong (the correct code appears also in the lines below) is
that 'latestPolledRequest' is always not null because the function
getLatestPolledRequest() returns a new object 'AttestRequest' in
case of request not found.

Then the check on result of 'latestPolledRequest' is removed, because
the function getLatestPolledRequest() always selects an attestation
request whose result field is null.