HisAppraiser: Fixed bug in writing field reportErrors
This patch fixes the generation of string to be written in
field reportErrors of table AuditLog. The string is obtained
from a list filled by HisReportValidator constructor, but the
generation process uses an object StringBuffer for all errors
except the first one.

The bug occurs when the list of errors contains at least 2
elements: in this case, the first error is copied to the
final string, but later it is overwritten by the concatenation
of subsequent errors.