HisAppraiser: Fixed bug in reports comparison
This patch fixes a bug in function comparePreviousReport() of
class HisReportValidator. The function is part of report validation
process and compares PCRs in the current report with PCRs of last
report received from the same host.

The function comparePreviousReport() generates an error message
when some PCR in the current report are not found in or differ from
PCRs in the previous report. The function didn't treat the case
of a PCR previously present and not found in the current report.