Installer: Added configuration properties for storing the IR on file
This patch introduces two configuration variables used by the
functions for reading and writing the report on file.

The property 'IR_DIR' permits the user to define the directory where
the integrity report will be stored; if not set (default) the IR
is stored on DB.

The property 'IR_DIGEST_METHOD' permits the user to chose the digest
method used to check the integrity of the IR; SHA-256 is its default
value.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>