HisAppraiser: Implemented functions for reading/writing IR on file
This patch adds functions for reading and writing IR on file instead
of storing it in table AuditLog. The reading function receives the ID
of the report on table AuditLog and the content of the corresponding
field Report. The function retrieves the report file, computes the hash
using the digest method read from the field Report and returns the XML
report string if the computed hash and the received one match.

The writing function receives the ID of the report on table AuditLog
and the report string; it computes the hash of the string using
the digest method read from configuration file, writes the report in
file and returns report string to be written on DB.

The new syntax of the field Report in table AuditLog is as follows:

	report_digest:<digest_method>:<digest_value>

It's still possible to save IR on DB. The writing function returns
the received reportString if the property 'IR_DIR' is not set; in
this case the field 'Report' contains the XML integirty report and
the reading function returns its content as is.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>