HisWebServices: Implemented new web service HisDownloadReportService
This patch adds a new web service (called HisDownloadReportService)
whose function fetchReport() returns the XML string of the integrity
report.

The introduction of this new web service is needed because,
since reports will be stored into the filesystem, the Portal
would have no means to obtain them if it is not running on
the same machine that executes the Appraiser tasks (currently,
the Portal connects directly to the DB).

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>