HisAppraiser: using readIR() and writeIR() to respectively read and write integrity reports
This patch replaces functions used by the Appraiser to read
and write the integrity report with readIR() and writeIR().

The benefit of using these functions is that callers
will not need to know how a report is stored in the
machine that executes Appraiser tasks but instead
they have only to pass the ID of the AuditLog record
and the report string as arguments.

A failure during writeIR() execution (e.g. given IR_DIR
does not exist) will result in the timeout of the attestation
request.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>