AttestationService: Added element "url" in attestation response
This patch adds to the attestation response a new element ("url")
that contains the URL of report on Portal.

The new structure of attestation result is as follows:
{
	"hosts": [{
		"host_name": "host1",
		"trust_lvl": "untrusted",
		"url": "http://server/OAT/report.php?id=34",
		"vtime": "2013-10-07T11:50:29.715+02:00"
	}]
}

The address of portal is read from a new property ("portal_address")
in OpenAttestationWebServices.properties; its default value is the
localhost hostname and it's set at installation time.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>