Installer, HisAppraiser: Added field reportCompareErrors
This patch introduces a distinction between errors related to
the report validation and errors related to the comparison with
last report received.
The first type of errors are still stored in the field
reportErrors of table AuditLog, while comparison-related
errors are stored in the new field compareReportErrors of
the same table.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>