Installer, HisAppraiser: Added field analysisResults to AttestRequest
This patch adds a new field (analysisResults) to table AttestRequest. It
will contain the result of analyses requested by the user and its syntax
is as follows:

	analysis_id|result|status|output_length|output;

where: "result" can be "true" or "false"; "status" contains the
status of the analysis execution; "output" is the output of the
analysis, if it exists.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>