HisAppraiser: Updated position of setResult() and setValidateTime()
This patch modifies the function validatePCRReport() so that
it will be compatible with the new extension that adds support
for multiple analyses.

With the new extension, first the overall verification result
is set to 'trusted' if the report validation is completed successfully.
Then, each analysis only sets the overall result to 'untrusted' if the
result of the analysis itself is negative. With this change, the
'result' field of an attestation response is obtained by the AND
(logic operator) of partial results of the analyses (and the result
of the report validation).

Finally, since there could be multiple analyses, the validate
time of an attestation response is determined after all analyses are
completed.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>