HisAppraiser: Updated behaviour of function validatePCRReport()
This patch updates the behaviour of function validatePCRReport()
by populating the new fields of the 'analysis_details'
subelement of the attestation result.
In particular, it stores all information in the analysisResult
field of the AttestRequest table, that will be parsed at the
time the attestation response is compiled.

With this patch, validatePCRReport() performs an analysis
on an integrity report called VALIDATE_PCR.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>