HisAppraiser: Implemented new OAT analysis COMPARE_REPORT
This patch adds a new function evaluateCompareReport() that
verifies if any errors occurred during report comparison. It
updates the fields Result and analysisResults of the received
entry of table AttestRequest and returns it.

The function evaluateCompareReport() replaces the check on
content of field reportCompareErrors in table AttestRequest.
In this way checking the result of the comparison of a subset
or all PCRs becomes an analysis called COMPARE_REPORT.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>