HisAppraiser: Fixed validation of Quote when PCR mask is not FFFFFF
The PcrSelect read from IR doesn't match the one used to obtain
DigestValue. In the first case the mask is read sequentially;
in the second case each byte is read in inverted order.

Then if the original mask is coded as follows:
	0 1 2 3 4 5 6 7 | 8 9 10 11 12 13 14 15 | 16 17 18 19 20 21 22 23 24
The mask to be used for DigestValue generation is as follows:
	7 6 5 4 3 2 1 0 | 15 14 13 12 11 10 9 8 | 24 23 22 21 20 19 18 17 16

This patch fixes this bug, reversing the order of bits of each
bytes composing the PCR mask.

Furthermore the amount of data to be verified is directly read
from the attribute ValueSize of element PcrComposite, as defined
by the TCG specification.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>