HisClient: Added XSD needed for including IML measurements in IR
This patch adds the XSD defined by TCG in order to include
in the report the SimpleObject element containing measurements
from the Integrity Measurement Log (IML). In particular, data is
retrieved from the securityfs interface of the Linux kernel.

These measurements, as the same as for OpenPTS, are grouped
by PCR so that it is possible to check whether the result
of the PCR extension operation (performed in software) with
the digests contained in a group (also called SnapshotCollection)
is equal to the corresponding PCR value retrieved from
the TPM quote.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>