HisAppraiser: Set new environment variables before calling analysis tool
This patch extends the set of environment variables set before
calling the external analysis tools. New variables are:
 - IR: The id of the report to be analysed
 - URL: The URL of HisDownloadReport web service

The analysis tool is expected to implement a SOAP client
using the given URL and ID to retrieve the integrity report.

In order to make the IR available to the web service, the
Hibernate transaction must be committed. This patch satisfies
this requirement by committing the transaction immediately after
after the AuditLog entry has been written, and before executing
the analyses.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>