HisAppraiser: Added field firstReport to table AuditLog
This patch adds a new field (firstReport) to the table AuditLog;
it stores the ID of the report containing the first part of the
list of measurements, that are measurements collected by the
machine executing the host agent since it has been started.

With the additional information, the Appraiser will be able
to build a report with the complete set of measurements
from partial reports (which contain data not already sent
to the Appraiser). The main benefit of this feature is
to save space on the Appraiser host because only the first
report will be large, while others should be significantly
smaller.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>