HisWebServices: Implemented reportType mechanism
This patch extends the class NonceSelect in order to notify the
Client to include or not in the integrity report the entire list
of measurements. If the property reportType is set to "continue",
the Client sends only measurements not already sent; otherwise
(reportType set to "start") it starts from the first measurement.

The setting of property reportType works as follows:
  - if the scalability is not enabled (default), the Client
    has to always send all the integrity measurements
    (reportType = "start");
  - if last report received from the Client is not valid,
    it is expected to send the entire list of IMA measurements
    (reportType = "start");
  - if last report received is valid and the scalability is enabled,
    the Client includes in the report only measurements not already
    sent (reportType = "continue").

The scalability mechanism works properly even if the Client is
rebooted, because the Host Agent always send the full report
when it is started.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>