HisAppraiser: Updated meas validation process with scalability support
This patch updates the measurements validation process, which
now handles partial reports.

For each SnapshotCollection, the current value of the PCR
is compared with the last one received. If they match, no
validation is needed; otherwise the StartHash value is
compared with the value of PCR read from last report. If
they match, the validation process continues as previously
defined; otherwise an error is reported.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>