HisClient: Implemented sending of a partial list of IMA measurements
This patch implements the client-side of the scalability mechanism.
The Client stores the number of bytes read from the measurement files
and last measurement read, so that it can ignore data previously sent
when the Appraiser requests a report of type 'continue'. Enabling
the scalability feature will considerably reduce the amount of data
included in each report.

Last byte read and last measurement sent are cleared each time
an error occurred while generating/sending an IR. In these cases
the Appraiser did not receive measurements, then the Client has
to restart from the first measurement.

The same behaviour is adopted when the Appraiser requested a report
type "start".

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>