HisAppraiser, HisWebServices, Portal: Updated function fetchReport()
This patch updates the function fetchReport() to reconstruct and
return the whole integrity report from partial ones when the
scalability mechanism is enabled.

The fetchReport() function accomplishes this task by retrieving
the reports referring to the same first report (audit_log records
having the same value for the firstReport column) in descending
order from the newest to the oldest. Then, after unmarshaling all
reports, it moves the data to the structure associated to the
newest one. Finally, it returns the string of the latest report
after performing the marshaling.

To maintain the same behaviour, fetchReport() always returns the
complete report, performing the reconstruction if necessary. It
is possible to retrieve also the partial report by adding the
'partial' parameter, set to true, to the function arguments.

The patch also updates the Portal and the HisDowloadReportService
in order to match the new behaviour of function fetchReport().

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>