HisAppraiser: Extended table AttestRequest for periodic requests support
This patch extends the table AttestRequest with four new fields:
threshold, expiration_time, last_read_time and current_processing_time.

The field threshold will contain the acceptable validity limit
requested by the user for periodic attestation requests. OAT
will guarantee to the user a result not older than the given
threshold.

The field expiration_time is the date when the periodic request
will be disabled. If null the request will never be removed.

The field last_read_time will contain the last time a periodic
request has been read; it will be used to automatically delete
requests whose result has not been read for a long time.

The field current_processing_time will contain the time needed
to validate the IR and execute the analysis; it will be used to
define how much a request must be anticipated so that OAT will
deliver a valid (fresh) attestation result to the user.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>