Installer, HisAppraiser: Added properties used by periodic req mechanism
Periodic request mechanism introduces two main issues: the user
wants to always have a valid result, and the Appraiser wants to
minimize the number of useless attestation requests sent to the
Client.

If a new attestation is submitted exactly when the previous one
validity expires, there will be no valid result until this
attestation ends. On the contrary, if OAT anticipates too much
the attestation in respect to the validity expiration, the workload
will increase with no advantage for the user.

Because each periodic attestation is identical, they are expected
to take approximately the same time, except for different workload
condition at attestation time. Then the duration of the attestation
is a good reference to define how much the next attestation request
must be anticipated in respect to the validity expiration.

The problem is represented by next diagram:

0sec                      10sec                                   130sec
|-------------------------|---------------------------------------|
T0                        T1         T2                           T3
|-------------------------|----------|----------------------------|
|-current_processing_time-|----------validity (120sec)------------|
                                     |-------anticipation (x)-----|

x = T3 - T2

x: how much time before previous request validity expiration OAT has
   to issue another attestation request?

If each periodic request requires exactly the same processing time,
the optimum anticipation (x) is that interval (T1 - T0). But
processing time depends from Appraiser and Client workloads, then
current attestation could require more time, resulting in an interval
with no valid result for the user.

For this reason, this patch introduces an anticipationFactor, that
will be multiplied for the current processing time in order to
introduce a tolerance to the variation of attestation time.

Final anticipation (x) formula is as follows:

x = current_processing_time * anticipationFactor

Then x depends from the current processing time and the system
administrator can modify the anticipationFactor in order to
guarantee more/less reliability to the user.
A higher anticipationFactor reduces the possibility that the
user won't obtain a valid attestation result, but it also
increases the Appraiser and Client workload.

However, it's necessary to set a limit to the to the anticipation
value (x), in order to limit the attestation request rate. That's
why this patch introduces a second property, called minAttestInterval.
It is a millisecond time value (default: 300000ms = 5min) and it
represents the minimum time interval between last result received
and next attestation request (T2 - T1).

Last property introduced by this patch is the maxIdleTime.
It is an interval (in days) after which a periodic request is
disabled if not read. As well as the anticipationFactor and
minRequestInterval, this is a method used to reduce OAT workload,
avoiding the sending of useless attestation requests.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>