HisAppraiser: Added to AttestDao the support for periodic requests
This patch extends the AttestDao with two functions related to
the periodic attestation requests mechanism.

The function getPendingRequests() returns a list of attestation
requests to be served by the given Client, including periodic
requests. It will be used by the polling WS to know if there
is a request to be served, and by the function submitReport()
to know which request to serve with the received report.
The function returns a list instead of a single request because
in this way it will be possible to introduce a more complex logic
in the process of choosing the request to be served.

The function disableUnusedRequests() disables requests whose
lastReadTime is older than configured maxIdleTime, or whose
expirationTime is up, or whose result is UN_TRUSTED. Then it
covers three out of the five causes of request deactivation,
previously introduced. Deletion by user and host unreachable
can't be managed by the Appraiser and will be covered by the
AttestationService.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>