AttestationService: Added new API for periodic attestation requests
This patch reintroduces in the AttestationService the API function
called "postHosts". This will be used to request asynchronous
attestations (periodic or not), obtaining an attestation ID to
be used to retrieve the results.

The format of a sample POST on the new postHosts API is as follows:
{
	"hosts": ["HOST_NAME_1", "HOST_NAME_2"],
	"analysisType":"ANALYSIS_NAME, ANALYSIS_PARAM;ANALYSIS_NAME",
	"timeThreshold":"50000",
	"expirationTime":"7d"
}

and the format of the corresponding response is as follows:
{
	"requestId":"REQUEST_ID"
}

This is also the format of a sample GET on the postHosts API, used to
retrieve the last valid attestation response, if it exists.
The user can also obtain last attestation result of an invalid
periodic task request; in this case the format of request is as
follows:
{
	"requestId":"REQUEST_ID",
	"lastResult": "true"
}

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>