HisAppraiser: Added constants used by the periodic attestation mechanism
This patch adds to the Appraiser some constants representing the
status of a periodic attestation request. The request can be stopped
for four reasons:
 - expirationTime is passed (value: -1)
 - maxIdleTime is reached (value: -2)
 - the used deleted the request (value: -3)
 - last attestation result was UN_TRUSTED (value: -4)
 - the host was unreachable for too much time (value: -5)

Reported values will be written in the field threshold of table
AttestRequest, that can also be null (asynchronous attestation
request), or positive (active periodic request).
OAT will use the value of threshold to determine the type and the
status of the attestation requests: if the value is negative, the
AttestationService will return an exception to the user, specifying
the cause of deactivation.