HisWebServices, HisAppraiser: Added support to periodic requests
This patch updates the HisPolling Web Service and the Appraiser in
order to support periodic attestation requests. Both components
now use the new function getPendingRequests() to obtain the
attestation request to be served, taking into account the presence
of periodic requests.

If the request to be served is periodic, the HisPolling Web Service
also updates the field requestTime. In this way it is possible
to known when last attestation request has been sent to the Client.

In case of periodic requests, the Appraiser skips analyses if reports
are identical, because launching the same analyses on the same set of
measurements will produce the same result. The field
currentProcessingTime is updated only when analyses are done, because
it is used to obtain the anticipation time and it must take into
account also the analyses duration.