HisAppraiser: Added functions createAuditLog() and createAlert()
This patch moves some functionalities of submitReport() in distinct
functions. The creation of the AuditLog entry has been moved to the
new function createAuditLog(), and the creation of an alert has been
moved to the new function createAlert().

This change does not alter the Appraiser behaviour, but it makes
clearer the next patches.