HisAppraiser: Introduced possibility of discarding identical IRs
This patch adds a new feature to the Appraiser, that can now
avoid saving the received IR if its PCRs are identical to the
ones of the last report received.

This feature is not activated by default and it can be enabled
using the previously introduced property DISCARD_IDENTICAL_IR
in file OAT.properties.