HisClient: Level of valuesType defined from PCR num instead of eventType
This patch updates the generation of the element valuesType. The level
(LV0 or LV1) was defined using eventType, but this method does not
cover all of eventTypes used by IMA. The eventType is now replaced
with the PCR number: valuesType for PCR 10 is always LV1; all other
PCRs have level LV0.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>