HisAppraiser: Skipped validation for PCRs not in the MLE pcrIMLMask
As part of the validation process, OAT extends all measurements in
the IR and compares the result with the value of PCR in the quote.
This approach guarantees that received measurements are correct,
but it could introduce problem with the PCRs of the Dynamic Root of
Trust and the ones extended by TrustedGRUB.

In order to disable the validation of critical PCRs, it was possible
to properly set the property PCR_SELECT, removing PCRs from IR and
then disabling their validation. The limit of this behaviour is that
the user can't obtain the values of that PCRs in order of analyse
them (e.g. compare them with the PCR_WHITE_LIST).

This patch solves the issue reading the value of new field
pcrIMLMask from table MLE and validating only PCRs specified
in the mask. In this way it is possible to obtain the complete list
of PCRs values even if the Client does not include all measurements
in the IR.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>