Installer, HisAppraiser: Added field pcrIMLMask to table MLE
This patch adds to the table MLE the new field pcrIMLMask,
that contains a bitmask of PCRs for which a host provides
a complete Integrity Measurement Log (IML). For PCRs
not included in the mask, OAT will not validate IR snapshots.

Adding this information is necessary because IR validation
fails for some systems (e.g. HP 8100) that do not provide
a correct IML for PCRs 0-7 or include software components
(e.g. TrustedGRUB) that extend PCRs without recording a log.
Also, the new mask is used to skip the validation for
resettable PCRs (17-23), not currently supported.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>