CommandTool, WLMService: Added element pcrIMLMask to the MLE API
This patch extends the MLE API with the new element pcrIMLMask,
whose value is written in the field pcrIMLMask of table MLE.

The new syntax for adding an MLE is as follows:
{
	"Name":"MLE_NAME",
	"Version":"MLE_VERSION",
	"OsName":"OS_NAME",
	"OsVersion":"OS_VERSION",
	"Attestation_Type": "PCR",
	"MLE_Type":"VMM/BIOS",
	"Description":"DESCRIPTION",
	"pcrIMLMask": "HEX_MASK"
}

If the new element pcrIMLMask is omitted, OAT sets the mask
to 000000, assuming that the host does not provide IML for
any PCR (this ensures compatibility with OAT 1.6).

The patch also updates the help message printed by the command
oat_mle, matching the new parameters.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>