Installer, HisAppraiser: Added field requiredPcrMask to AnalysisTypes
This patch adds the field requiredPcrMask to the table AnalysisTypes.
It will contain the mask of PCRs required by the analysis, which
will be compared to the pcrIMLMask in order to verify if the analysis
can be launched for the requested host.
If the analysis requires PCRs not included in the pcrIMLMask, it
won't be launched: the analysis result will be "false" and an
error message will be returned to the user.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>