HisAppraiser: Analysis skipped if IR doesn't contain required measures
This patch enhances the external analysis mechanism skipping the
analysis if it requires measurements not included in the
pcrIMLMask defined for the host's MLE.

Each analysis request can contain a parameter whose syntax is
as follows:

	pcrs=#|#|#|#

where each "#" is the number of a PCR.
If the parameter is not specified, the external analysis tool is
expected to use all PCRs defined in the requiredPcrMask.

The function analysisLauncher() checks if requested PCRs are in
the requiredPcrMask defined for the analysis; if so, the function
checks if measurements for each PCR are provided by the host.
If all checks succeed, the analysis is launched, otherwise an
OAT_ERROR is returned to the user.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>