CommandTool, WLMService: Added element requiredPcrMask to analysisType API
This patch adds the element requiredPcrMask to the analysisTypes API.
It permits to set the value of the element requiredPcrMask in
table AnalysisTypes.

The new syntax for adding an analysis is as follows:
{
	"name": "ANALYSIS_NAME",
	"module": "MODULE_NAME",
	"version": VERSION_NUMBER,
	"url": "ANALYSIS_URL",
	"requiredPcrMask": "HEX_MASK"
}

If the new element requiredPcrMask is omitted, OAT sets the mask
to 000000, assuming that the analysis does not need IMLs for any PCR.

The patch also updates help message printed by the command
oat_analysisType, adding the new parameter to the example request.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>