CommandTool: Added command for conversion of PCRs list into hex bitmask
This patch adds to the CommandTool a new script, called
oat_pcrListToMask. The new command receives a list of PCRs and
returns the corresponding hexadecimal mask.
It has been introduced in order to simplify the configuration of
new analyses and the host MLE (that requires specifying respectively
the elements requiredPcrMask and pcrIMLMask).

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>