HisAppraiser, HisWebServices: Updated scalability mechanism
This patch updates the scalability mechanism in order to solve
a problem introduced by the use of pcrIMLMask. If the mask changes,
the chain of IRs could not be reliable because some measurement
on a previous report may have not been verified. In particular,
there may be the case where a malicious client could have included
in a previous IR a corrupted snapshot corresponding to a PCR not in
pcrIMLMask (thus, not validated) and, after the OAT administrator
enabled that PCR, the validation may succeed if the snapshot of
the currently received IR is correct.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>