HisAppraiser: Skipped measurements validation if not needed
This patch fixes a bug in the constructor of class HisReportValidator.
The class is used in two way, validating current report and just
parsing the received report string.
In the second case, the constructor receives only the report string
and the validation of the measurements must be skipped. Otherwise,
since the machineName is missing it's not possible to obtain the
pcrIMLMask and an exception is raised.

Signed-off-by: Nicola Barresi <nicola.barresi@polito.it>
Signed-off-by: Roberto Sassu <roberto.sassu@polito.it>